adds bitfield class (has bug)
