+++ owltorrent-network/src/main/java/edu/rice/owltorrent/network/messages/BitfieldMessage.java
@@ -1,5 +1,6 @@
 package edu.rice.owltorrent.network.messages;
 
+import edu.rice.owltorrent.common.entity.Bitfield;
 import edu.rice.owltorrent.network.PeerMessage;
 import java.nio.ByteBuffer;
 import java.util.BitSet;
@@ -12,9 +13,9 @@ import java.util.BitSet;
 public class BitfieldMessage extends PeerMessage {
 
   public static final int BASE_SIZE = 1;
-  public BitSet bitfield;
+  public Bitfield bitfield;
 
-  public BitfieldMessage(BitSet bitfield) {
+  public BitfieldMessage(Bitfield bitfield) {
     super(MessageType.BITFIELD);
     this.bitfield = bitfield;
   }
@@ -30,18 +31,9 @@ public class BitfieldMessage extends PeerMessage {
     return buffer.array();
   }
 
-  public boolean getBit(int index) {
-    /*
-     The order in the BitSet is 0 at the least significant digit and 7 at the most significant digit for the first byte.
-     Therefore, a conversion is necessary to make the most significant bit be at index 0.
-     For index i, the sum of i and its actual index in the BitSet is ((i/8) * 2 + 1) * 8 - 1.
-    */
-    int actualIndex = 8 * ((index / 8) * 2 + 1) - 1 - index;
-    return this.bitfield.get(actualIndex);
-  }
-
   public static BitfieldMessage parse(ByteBuffer buffer) {
-    BitSet bitfield = BitSet.valueOf(buffer);
+    BitSet bitSet = BitSet.valueOf(buffer);
+    Bitfield bitfield = new Bitfield(bitSet);
     return new BitfieldMessage(bitfield);
   }
 }
